pyproject.toml
[tool.poetry]
name = "nextgenslm"
version = "0.1.0"
description = ""
authors = ["Marc Green <marc1701@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
SciPy = "^1.11.4"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
seaborn = "^0.13.1"
torchaudio = "^2.1.2"
openpyxl = "^3.1.2"
soundfile = "^0.12.1"
librosa = "^0.10.1"
pytorch-lightning = "^2.1.3"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
